bls_data_sync/src/requirements.txt
boto3==1.34.69
requests==2.31.0
python-dotenv==1.0.1
pyspark==3.5.0 